% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_sites.R
\name{extract_buffer_data}
\alias{extract_buffer_data}
\title{Extract raster values within buffer zone}
\usage{
extract_buffer_data(raster_data, buffer_zone)
}
\arguments{
\item{raster_data}{SpatRaster object}

\item{buffer_zone}{sf buffer zone object}
}
\value{
Data frame with extracted values and coordinates
}
\description{
Extract raster values within buffer zone
}
\keyword{internal}
